+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 
 2.3.0
 +++++
+* `az acs/aks install-cli` will install to under %USERPROFILE%\.azure-kubectl on Windows
 * `az aks install-connector` will now detect if the cluster has RBAC and configure ACI Connector appropriately
 * Create role assignment to the subnet when it's provided.
 * Add new option "skip role assignment" for subnet when it's provided